+++ backend/main.cpp
@@ -1,0 +1,26 @@
+#include "stdafx.h"
+#include "wood_test.h" // test
+
+int main(int argc, char **argv)
+{
+
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	// GoogleTest
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	// wood_test::run_all_tests();
+
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	// Display
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	wood_globals::DISTANCE = 0.1;
+	wood_globals::DISTANCE_SQUARED = 0.01;
+
+	opengl_globals::shader_type_0default_1transparent_2shaded_3shadedwireframe_4wireframe_5normals_6explode = 2;
+	opengl_globals::shaders_folder = "C:\\IBOIS57\\_Code\\Software\\Python\\compas_wood\\backend\\src\\viewer\\shaders\\";
+	opengl_render::render(wood_test::type_plates_name_vda_floor_0);
+
+	// bool type_plates_name_cross_square_reciprocal_two_sides();
+	// bool type_plates_name_cross_square_reciprocal_iseya();
+	// bool type_plates_name_cross_ibois_pavilion();
+	return 0;
+}